+++ changelog.md
@@ -1,2 +1,7 @@
 # 🆕 Features
+
 - Replace the sneak key with a custom key(⚠️Singleplayer only)
+
+# 🐞 Bug fixes
+
+- The block will be broken with empty hand while the hand stack is empty